ExternalRDD.apply() set true into nullable if T in RDD[T] is non-primitive type
