+++ sql/core/src/test/scala/org/apache/spark/sql/DataFrameSuite.scala
@@ -1722,15 +1722,4 @@ class DataFrameSuite extends QueryTest with SharedSQLContext {
         "Cannot have map type columns in DataFrame which calls set operations"))
     }
   }
-
-  test("SPARK-19959: df[java.lang.Long].collect includes null throws NullPointerException") {
-    val dfInt = sparkContext.parallelize(Seq[java.lang.Integer](0, null, 2), 1).toDF
-    assert(dfInt.collect === Array(Row(0), Row(null), Row(2)))
-    val dfLong = sparkContext.parallelize(Seq[java.lang.Long](0L, null, 2L), 1).toDF
-    assert(dfLong.collect === Array(Row(0L), Row(null), Row(2L)))
-    val dfFloat = sparkContext.parallelize(Seq[java.lang.Float](0.0F, null, 2.0F), 1).toDF
-    assert(dfFloat.collect === Array(Row(0.0F), Row(null), Row(2.0F)))
-    val dfDouble = sparkContext.parallelize(Seq[java.lang.Double](0.0D, null, 2.0D), 1).toDF
-    assert(dfDouble.collect === Array(Row(0.0D), Row(null), Row(2.0D)))
-  }
 }
